visual: extract ESIL from vmenus.c to esil.c
